Ameliorate displaying the router's state
'Plugged' bit is currently used to check if the router is
in 'uninitialized unplugged' state or not.
This bit gets set to '1' only when lane initialization completes.
The bit to be checked should be 'router detected', but since
router logically exists in the system, remove checking for this
state entirely.

Signed-off-by: Rajat Khandelwal <rajat.khandelwal@intel.com>